refactor: 冗長な抽象化部分の削除
