--- 7_Clima-iOS13/Clima/API/IcanhazdadjokeAPI/IcanhazdadjokeAPIRequest.swift
@@ -1,12 +1,0 @@
-//
-//  IcanhazdadjokeAPIRequest.swift
-//  Clima
-//
-//  Created by HIROKI IKEUCHI on 2024/10/04.
-//  Copyright © 2024 App Brewery. All rights reserved.
-//
-
-import Foundation
-
-// 名前空間の定義
-enum IcanhazdadjokeAPIRequest {}
