+++ 7_Clima-iOS13/Clima/API/Models/IcanhazdadjokeData.swift
