+++ 7_Clima-iOS13/Clima/API/Service/APIService.swift
@@ -1,8 +1,8 @@
 //
-//  APIServiceProtocol.swift
+//  APIService.swift
 //  Clima
 //
-//  Created by HIROKI IKEUCHI on 2024/10/04.
+//  Created by HIROKI IKEUCHI on 2024/10/11.
 //  Copyright © 2024 App Brewery. All rights reserved.
 //
 
@@ -10,15 +10,12 @@ import Foundation
 import HTTPTypes
 import HTTPTypesFoundation
 
-protocol APIServiceProtocol {
-    associatedtype APIError: Error, Codable
-}
-
-// MARK: - 共通機能
-
-extension APIServiceProtocol {
-    func request<Request>(with request: Request) async throws(APIServiceError<APIError>) ->
-    Request.Response where Request: APIRequestProtocol {                
+final actor APIService {
+    
+    static let shared = APIService()
+    
+    private func request<Request>(with request: Request) async throws(APIServiceError) ->
+    Request.Response where Request: APIRequestProtocol {
         // リクエストの送信
         guard let httpRequest = request.buildHTTPRequest() else {
             throw APIServiceError.invalidRequest
@@ -31,19 +28,10 @@ extension APIServiceProtocol {
         }
         
         // レスポンスへのハンドリング
-#if DEBUG
-        let _responseMessage = String(data: data, encoding: .utf8) ?? "" // swiftlint:disable:this identifier_name
-        print(_responseMessage)
-#endif
         if !(200..<300).contains(response.status.code) {
             // レスポンスのコードがエラー範囲の場合
-            let apiError: APIError
-            do {
-                apiError = try JSONDecoder().decode(APIError.self, from: data)
-            } catch {
-                throw APIServiceError.apiResponseParseError(error)
-            }
-            throw APIServiceError.apiError(apiError)
+            let reason = String(data: data, encoding: .utf8) ?? ""
+            throw APIServiceError.apiError(reason)
         }
         
         // レスポンスをデータモデルへデコード
@@ -53,5 +41,23 @@ extension APIServiceProtocol {
         } catch {
             throw APIServiceError.apiResponseParseError(error)
         }
+    }
+}
+
+// MARK: - Call API
+
+extension APIService {
+    /// 天気の取得
+    func fetchWeather(for location: WeatherAPIRequest.Location) async throws(APIServiceError) -> WeatherData {
+        let request = WeatherAPIRequest(location: location)
+        let weatherData = try await self.request(with: request)
+        return weatherData
+    }
+    
+    /// jokeの取得
+    func fetchRandomJoke() async throws(APIServiceError) -> IcanhazdadjokeData {
+        let request = IcanhazdadjokeAPIRequest()
+        let joke = try await self.request(with: request)
+        return joke
     }
 }
